rename rule tested-class-name-in-top-level-describe
