+++ lib/rules/tested-class-name-in-top-level-describe.ts
